+++ lua/vv-explorer/actions/init.lua
@@ -5,6 +5,8 @@ local H = require('vv-explorer.actions.helpers')
 local M = {}
 
 M.node_under_cursor = H.node_under_cursor
+M.find_row = H.find_row
+M.expand_to_file = H.expand_to_file
 
 require('vv-explorer.actions.navigation').attach(M, H)
 require('vv-explorer.actions.filter').attach(M, H)